+++ plugins/ucc_chat/lib/ucc_chat/models/notification_setting.ex
@@ -3,6 +3,8 @@ defmodule UccChat.NotificationSetting do
 
   use UcxUccWeb.Gettext
 
+  alias UccChatWeb.Admin.Page.ChatGeneral
+
   @audio_options [
     {"none", "None"},
     {"system_default", "Use account preferences (Default)"},
@@ -28,21 +30,21 @@ defmodule UccChat.NotificationSetting do
   def options(:audio),        do: @audio_options
 
   def options(:audio_mode), do: [
-    {"default", gettext("Default (%{default})", default: "Mentions")},
+    {"default", gettext("Default (%{default})", default: get_system_audio_name())},
     {"all", ~g(All messages)},
     {"mentions", ~g(Mentions)}
   ]
 
-
   def options(field) when field in ~w(desktop mobile)a, do: [
+    {"default", gettext("Default (%{default})", default: get_system_name())},
     {"all", ~g"All messages"},
-    {"mentions", ~g"Mentions (default)"},
-    {"nothing", ~g"Nothing"}
+    {"mentions", ~g"Mentions"},
+    {"none", ~g"Nothing"}
   ]
 
   def options(:email), do: [
     {"all", ~g"All messages"},
-    {"nothing", ~g"Nothing"},
+    {"none", ~g"Nothing"},
     {"preferences", ~g"Use account preference"}
   ]
 
@@ -52,4 +54,13 @@ defmodule UccChat.NotificationSetting do
     {"preferences", ~g"Use account preference"}
   ]
 
+  def get_system_name() do
+    ChatGeneral.lookup_option(:notifications,
+      UccSettings.desktop_notifications_default_alert)
+  end
+
+  def get_system_audio_name() do
+    ChatGeneral.lookup_option(:notifications,
+      UccSettings.audio_notifications_default_alert)
+  end
 end