+++ plugins/ucc_chat/lib/ucc_chat/models/schema/notification_setting.ex
@@ -4,7 +4,7 @@ defmodule UccChat.Schema.NotificationSetting do
   embedded_schema do
     field :audio_mode, :string, default: "default"
     field :audio, :string, default: "system_default"
-    field :desktop, :string, default: "mentions"
+    field :desktop, :string, default: "default"
     field :duration, :integer, default: nil
     field :mobile, :string, default: "mentions"
     field :email, :string, default: "preferences"